refactor(classtable): split ClassTableViewModel across three files
Same treatment as AppState: behaviour moves into extensions in sibling
files, stored properties stay on the class because Swift extensions
cannot hold them.

    ClassTableViewModel.swift          996 -> 388
    ClassTableViewModel+Grid.swift            246
    ClassTableViewModel+Editing.swift         242
    ClassTableViewModel+Loading.swift         120

Grid is the 衝堂 work — contiguous blocks, transitive conflict clusters,
and routing a tap on a shared cell through the picker instead of guessing
which course was meant. Editing is add/delete/rename/recolour, kept
together because each has to persist locally *and* tell the backend.
Loading is fetch plus the dataDidUpdate broadcast that wakes Home and the
Live Activity coordinator.

17 members widened from private to internal for the crossing, same
trade-off as the AppState split.

All three go on the macOS exclusion list next to ClassTableViewModel
itself, which is already there — the Mac has its own MacClassTableView
and does not compile the UIKit-dependent timetable at all. The guard
caught them as unassigned before this was committed, which is what it
is for.

iOS and macOS both build. The 12 failing unit tests are the pre-existing
ones this branch surfaced by making the test bundle compile again. One of
those files does construct a ClassTableViewModel, but only in
classTableViewModel_displayLabel_formatsSemesterCode, which passes — and
displayLabel stayed in the main file regardless.
